Redirect when selected_account not found [#87201]
If a :selected_account param exists, and that account is not among the accounts
that need reconciling for the current facility, redirect the user back to the
"index" page for PO or CC accounts to be reconciled. This will either select the
first account, or display a notice that there are no accounts that need
reconciliation.

Previously this situation would error out when #get_unreconciled_details
called OrderDetail.account_unreconciled with @selected being nil.