[#60971 60973] Add more message types to the Notices tab
This takes the existing "Notices" tab in the header and expands it to include:

* Disputed Orders
* Problem Orders
* Problem Reservations
* Training Requests

The items (and total count) should reflect only the messages appropriate for
the current user in the current facility context. That is, if the user does not
have the Ability to :manage training requests, they will not appear in the
notifications menu even if some exist.

If there are no messages for a given category, it is not included in the
drop-down. If there are no messages of any type for the user and context, the
"Notices" tab should not appear at all.

Each menu item is a link to a page to view and manage the appropriate type.
"Disputed Orders" is new, and is under the Billing tab. It should show all
disputed order types, both reservation and non-reservation.

It includes:

* Extraction of the OrderStatus factory
* Addition of specs for the TransactionHistoryHelper and facilities helpers
* Disambiguation of "state" in the in_dispute scope[*]
* Making #current_ability a public method in the ApplicationController

[*] It was possible to get an ambiguous column error for "STATE" when using
TransactionSearch, and this should avoid that.

Closes #326