[#98879] Accessory popup error in New/In Process Queue
When ending a reservation from the new/in process queue, the accessory
window was not popping up properly. The `#pick_accessories_dialog`
needs a couple classes (modal, fade, etc) in order to display properly.
Without explicitly declaring it on the page, the javascript will create
the element with all the necessary attributes.