[#60971 #60973] Add MessageNotifier class
This determines what message summary the current user may see
in a facility context (or no facility context).

For example, if there are problem reservations in Facility A, they will not be
part of the count when under Facility B, and the count will be zero when scoped to
no facility, as on the Home page.