[#119006] Optimize journal closing
NU has had timeout issues with closing journals with many order details.

This changes to use `update_all` on order details on successful journal
closing instead of an update on each individual order detail. We lose
the `versions` audit details, but the performance is about an order of
magnitude faster. After this, we should be able to drop the timeouts we
increased back to a normal level.

Benchmarking code can be seen in 445af028f94dfc70927653a0915fcaa38237ff31

100 items in a journal:

```
original
  0.110000   0.030000   0.140000 (  0.173225)
  0.070000   0.000000   0.070000 (  0.101981)
  0.070000   0.010000   0.080000 (  0.102546)
  0.060000   0.010000   0.070000 (  0.102396)
  0.070000   0.000000   0.070000 (  0.115820)
  0.060000   0.010000   0.070000 (  0.099806)
  0.060000   0.010000   0.070000 (  0.100836)
  0.070000   0.000000   0.070000 (  0.102700)
  0.090000   0.020000   0.110000 (  0.130646)
  0.070000   0.000000   0.070000 (  0.097616)

update_all
  0.030000   0.000000   0.030000 (  0.041834)
  0.010000   0.000000   0.010000 (  0.018699)
  0.010000   0.010000   0.020000 (  0.019488)
  0.020000   0.000000   0.020000 (  0.023885)
  0.020000   0.000000   0.020000 (  0.029898)
  0.020000   0.000000   0.020000 (  0.026476)
  0.020000   0.010000   0.030000 (  0.026347)
  0.020000   0.000000   0.020000 (  0.027293)
  0.010000   0.000000   0.010000 (  0.024198)
  0.020000   0.000000   0.020000 (  0.027155)
```

Closes #343