[#60973] Update Facility staff-related permissions
This should bring the Notices drop-down permissions in line with what users may
access through other parts of the UI.

The notices drop-down was using Ability permissions to determine what to show
the user, while the corresponding links in the UI were not using Ability checks.
So we ended up with inconsistencies, like hiding the problem reservations tab
from Staff and Senior Staff, while showing problem reservations in the notices
drop-down.

This came down to replacing the :manage (anything goes) permissions to
Reservation and Order objects for Staff (that is, #operator_of?(current_facility))
with specific grants, and using these Ability checks in the UI to match the
behavior of Notices.

Closes #359