Clean up owner_user association
We had a test that would fail about 20% of the time because the
`owner_user` association was losing the `user_role` conditional.

Calling `Account.first.owner_user` would result in the SQL:
`SELECT `users`.* FROM `users` INNER JOIN `account_users` ON
`users`.`id` = `account_users`.`user_id` WHERE
`account_users`.`account_id` = 204 AND `account_users`.`deleted_at` IS
NULL LIMIT 1`. If I flipped the order of the keys in the hash, I would
get the “Owner” condition, but not the deleted_at.

This seems like it could be a bug in ActiveRecord, but since we’re
still on 3.2, so I didn’t want to delve too far down.