Create an initial migration so db:migrate works
This is part of my slow progress to get a functional Oracle Docker image running. Oracle has a limit on the length of index names, so a straight `db:schema:load` would not work.

And, when NUCore was brought in to git, there was a pre-existing schema which was not captured in the migrations, thus `rake db:migrate` would not work from a blank database.

I tried two techniques in order to get `rake db:migrate` working and
decided on this one.

The first:
* Find the first commit’s schema.rb and put that in an initial migration.
* Delete all migrations up to that point
* Fix any issues caused by missing data

That was fairly simple, but it lost much of the history of the project,
so I tried another process.

* Find the first commit's schema.rb and put that in an intial migration.
* Run `rake db:drop db:create db:migrate` and as I found errors, correct
the original migration accordingly.
* Fix issues caused by missing data or API changes in the migrations

They are similar, but I'm happier with this result.

* It gives a better history of the project
* Index names are more in line with Rails standards

Notes on the changes:
* Some columns have changed ordering within the tables -- not an issue
for new deploys, but could result in weirdness on existing development
commits when working off of existing databases
* Timestamps for some tables now have `null: false`.
* The new version of foreigner uses 1.9 hash syntax, so `add_foreign_key`
in `schema.rb` will be using the new syntax while the rest of schema.rb
uses the `=>` syntax.

Remove references to nucs_ in migrations

Closes #377
More details in the PR