Remove #operator? and #manager? methods from Role
We removed #manager? as it was not used anywhere and removed #operator? as the
role doesn't really make sense except in the context of a Facility.

The #operator_of?(facility) and #manager_of?(facility) methods remain.

Removing #operator? meant reworking Abilities for Notifications:
* Facility Operators should be able to read notifications
* If a user has a Notification, the user should be able to read it

This also removes the concept of an app_admin_name in favor of using app_name
everywhere.

Closes #375