More powerful rake task for clearing oracle databases
I’ve had bad luck with `rake db:drop db:create` on oracle. We had
`db:oracle_drop`, but that parsed the schema.rb file for table names,
so when dealing with a database that didn’t match the schema.rb, it
might not clear everything out. This drops every table and sequence. So
it’s very dangerous, hence why it’s only allowed to run in dev & test
modes, and while it’s suffixed with _severe.