Fix migration from scratch issue
When running the migration for a new project, the `change_column
:products, :facility_account_id, :integer, null: false` was failing
because there was a foreign key attached to it. Oddly, it was only
failing when run on the stage server database; it worked fine locally.

This changes the order of operations so the index doesn’t exist until
the column is in the state it needs to be. It also changes from
explicit sql to `add_foreign_key`.