[#98883] Add Account Manager user role
This adds a new global user role called Account Manager, which has access to
the User administration tab in a cross-facility context, which is something
that no other roles (even global admins) have.

Previously many of the navigation tabs and links used Role checks to determine
display. The changes here should all use Ability checks instead. This does mean
we are now explicitly taking away abilities from the global administrator role,
but it shouldn't change the behavior this role has seen all along. For example,
the cross-facility Billing tab had a condition to make it visible only to
Billing Administrators, which also meant it was not visible to global admins.
Converting this to an Ability check meant removing permissions from global
admins, which until now had an across-the-board can(:manage, :all). The Ability
restrictions were also the reason behind the spec change in 29bfcd0;
it's something that probably should have been tested as a Billing Administrator
all along.

This change also:

* Refactors the global admin tabs to use the NavTab presenters from #383
* Removes the unused Role.grant_role! method
* Adds FacilityAccountOrdersController to replace the "recent" case in
  FacilityAccountsController#show_statement
* Adds an OrderDetailPresenter, which should help reduce the size of the
  OrderDetail model and helper methods in general over time
* Sets a facility for CC accounts in the demo seed

This change does not include a UI for granting this new role to users.

Closes #397