[#98883] Add UI for managing global user roles
This adds a UI, available only to Global Administrators, for granting and
revoking global user roles, based on the one in place for facility user roles.

A user may not grant or revoke roles for itself. It would be easy to lock
yourself out, and would complicate the controller actions.

A user may have multiple global roles, though it might result in some odd
behavior. For example, the Global Administrator and Account Manager have some
conflicting abilities; Account Managers have access to the cross-facility Users
tab that Global Administrators do not. The functions of both roles should work,
though it may be difficult or impossible to reach them through navigation.

This change adds some validations on UserRole that might make existing data
technically invalid, in that we have some duplicate UserRoles and nonexistent
role names in production. We are not expecting this to be a problem since
granting/removing roles always involves removing all global roles first, valid
or invalid.

Closes #407