Add Strong Params to Facilities
I’m going to need to protect a couple columns on Facility for the
credit card processing, so I figured I’d bring in strong_params.
Getting it across the board will be a big undertaking (probably will do
it with the Rails 4 upgrade), but this starts laying the groundwork.