Clean up #show_statement action
This removes the '.all' in the @statements chain. Unsure why this was being
converted into an Array. It's now an AR relation.

This also removes the explicit '.order' from the @statements chain as it's
redundant; Statements already has an ordered default scope.

The assignment of @facility remains though feels obsolete; views should be able
to use current_facility instead.