[#121326] Send emails for training requests
Admins can set a comma-separated list of email addresses who will receive
notification anytime a user submits a training request. If the field is left
blank, no emails will be sent. We do very simple validation against the emails.

This is on a per-instrument basis. We decided against any facility-level or
mass configuration.

Closes #425