[#121871] Printable order detail view
UIC requested the ability to print the details of an order detail. We had previously
removed the order detail edit page in favor of the edit modal, however this did not
work well for printing. This change makes the modal also available as a regular
page which is accessible by right clicking the link or from the "Needs Approval"
email.

* Upgrade clockpunch to make non-modal order detail edit form work
  Relevant change: https://github.com/tablexi/clockpunch/pull/1
* Make order detail edit page printable
* Link to printable order detail manage page
* Update links from needs approval email

Closes #430