[#102132] Update new reservation account select
This updates and cleans up the partial used for selecting accounts on the new
reservation form.

The default should be no payment source selected, if there are multiple accounts
available. If the user selects a payment source but there is an error on create,
that payment source should remain selected.

If only one account is available, that single account is selected automatically.

Closes #441