[#120455] Split accounts reports
Introduces a reports querier and transformer service object.
The querier has an interface for finding order details.
And the transformer takes a collection of order details and
returns a modified collection of order details. We added the
transformer logic to generate virtual "split" order details
to pass to the report objects.

Closes #456.