Improve performance of Account General Report
This has negligible performance implications in open, but in NU where
`to_s` does a query to pmu_department, this saves quite a bit of time
when many OrderDetails use the same account. Testing against 1000 order
details on the same account, it cuts the load time from ~3.2 seconds
down to ~2.5 due to the savings of N+1 queries.

This really only works because the report view itself will call `to_s`
implicitly when outputting the account as the first column, but it does
improve performance, so I think it’s worth it.