[#115285123] Send order notifications to facilities
This adds order_notification_recipient to the Facility model.

If given a value, this address will receive order notifications.
These are generated the same time as receipts, and use the same templates,
with minor differences: the Subject line is different, and only receipts should
say "Thank you for your order".

Closes #515