[#122397] Prevent multiple reservations on the same order detail
It is possible for two reservations to be created for the same user if
they double click the “Create” button in quick succession and the
transactions overlap. This only happens on a multi-threaded server like
unicorn.

This also cleans up duplicates and adds a unique constraint on
`reservations.order_detail_id`. All duplicates I found in practice are
exact duplicates.

Also related to NU #124969, UIC #122397.