[#116373301] Add Projects index and new/create (#567)
This adds a new facility admin tab for viewing lists of Projects and creating new ones.

It includes some changes to NavTab to make it possible to change the tabs seen in an admin context by altering the LinkCollection.tab_methods array.

It also now treats including "projects" in the Gemfile as a feature flag. If 'projects' is in the gemfile, the engine is active.