Fix OrderStatus table on migration from scratch
When running `rake db:migrate db:seed` from scratch, the migrations would
result in improperly ordered OrderStatuses.

Adding the Settings.yml's order statuses based on the `order_details.status_change_hooks`
we listen to as observers happens before db:seed, as is the reconciled status. This
will now add them as part of `db:seed`.

I'd like to see if we can remove the `ActiveRecord::Observer` as part of the
Rails 4 upgrade since that is moved out into a separate gem, but that may not
be worth the effort.