Isolate namespace
This is the default configuration, but caused problems when loading a sanger page
using the default layout. This puts all of the calls to main app routes as calls
to `main_app`. The ones that didn't need it are called from Helpers, which get
loaded because the controllers are inheriting from `::ApplicationController`.

When we get to the admin side, there will likely be more routes
that need to be given the `main_app` prefix.