Upgrade/replace jquery-ui
jquery-ui-sass-rails is no longer maintained. Switch to using
jquery-ui-rails instead.

jquery-ui-rails uses sprockets `require` syntax, and SASS is not able
to do the dependent requires when called with `@import` so we need to
explicitly require each individual component that we need.

I went through the whole list of components, and I believe these are
the only ones we need.