[#126687] AM/PM and Date changes did not trigger pricing update
See https://github.com/tablexi/nucore-open/pull/155 for where this
broke. We were trying to fix a 500 error when the ajax was triggering
with an empty quantity. However, this then ignored the call for AM/PM
and date changes.

This now just makes sure there is at least something in the field. If
you do something silly like type in a non-numeric into quantity, the
quantity and pricing changes get thrown away by the backend (see the
change in PR #155).