[#126788] Fix where_ids_in problem when ids is empty for oracle (#166)
If ids is empty, then it resolves to `where(nil)` in Rails 4 or `where("")` in
our old Rails 3.2 code, which would load _every_ order detail. This was causing
an error on the Export as CSV report generation because there would be a blank
account.

This should be pulled up into open.