[#120869939] Download Sanger Sequencing template file (#678)
This modifies the batch creation so that reserved wells in unused rows don't get persisted: they appear as if they're blank in the well-plate export. But they still show up as "reserved" when creating the well plate.

It also adds a ActionController::Renderer for CSV to the root app so we can do render csv: ObjectRespondingToToCSV.new, filename: "somefilename". There are a few places elsewhere in the app I'd like to swap this in in place of CSVHelper, but that seemed outside the scope of this PR.

There are a few fields in the CSV that I'm not really sure if they're supposed to be dynamic or not. I figure we can address those when we meet with Chris next week.