[#126388] Allow an instrument to be marked offline/down (#677)
This adds OfflineReservations and a CRUD UI, temporarily restricted to global administrators.

Reservation is now an STI model, with the base Reservation (with type nil) used
as it was before: a reservation placed by or for a user. The other types are AdminReservation
(which existed before, without STI) and OfflineReservation, which is new with this change.

This change also removes the before_validation hook to run set_all_split_times for Reservations.
Since a large number of specs relied on the callback, this introduces a flag,
Reservation#split_times, to restore that behavior.