[#126393] Support offline reservation categories (#711)
This adds a "category" field to the Reservation model.

With this version, a category value may exist only for OfflineReservations,
and may be one of three hardcoded values: operator_unavailable,
out_of_order and scheduled_maintenance. It may also be null.

A user creating or editing an offline reservation through the UI may choose
one of these values or none.