Attempt 2: Prevent race conditions when debouncing email sending (#720)
The major change here is that we will simply bail in the case of
getting a record not unique error. We also don’t need to lock the table
in this case.

In the case where the uniqueness validation becomes invalid in the race
condition, then `update_attributes` will fail, so we won’t yield and
bail instead.

I actually tested this code on the staging server.