[#126395] Add admin and offline reservation (downtime) reporting (#724)
This adds a third kind of instrument report for instrument-not-available
due to offline reservations or admin reservations. It's handled in
Reports::InstrumentUnavailableReportsController; the previous two types are in
Reports::InstrumentReportsController & Reports::InstrumentDayReportsController.

The data comes out of Reports::InstrumentUnavailableReport though the specs are
at the controller level to ensure it's following the interface that
Reports::ReportsController and its templates expect.

It also includes the addition of a "duration" transient attribute to the
Reservation factory.