[#130559] Exclude non-purchased "add-to" orders from export raw
Prevent order details that were “added to” another order by the
facility staff, but were not complete from appearing in export raw.
These would be instrument orders that don’t have a reservation or
service orders where the order form has not been uploaded.

This takes care of it in two ways, either one would work on its own:
* Remove the `ordered_at` from the merge order. Now the querier won’t
find an ordered_at to search.
* Force the query to only return purchased orders. Once the merge order
is ready, the order details are moved onto the main order and then
destroyed.

The only thing I’d like to figure out is how to have it keep a null
order status until the purchase is complete (this is how carts work).
This seemed more tricky than it was worth in dealing with the
OrderDetailObserver.