[#127094] Normal users could place orders in the past (#765)
The big problem is that the account assignment on order_detail was autosaving the reservation without enforcing the non-admin validations (save_as_user).

The first idea I tried was doing update_columns so callbacks didn't get triggered, but it turns out we do need validations.

Next idea was flipping save and save_with_extended_validations so the extended validations (in the future and within 14 day window) are enforced by default, but not when purchasing as an admin. However, too many of our tests rely on those validations not being enforced as part of setup.

This refactor makes sure that reservation is not built until the last moment possible.