New account needs to be in same price group for UIC
When we change accounts, in order to make sure we end up with valid
price policies, the new account needs to be in the same price group as
the old one. In open/nu it’s fine because of user price group
memberships, but in UIC, it’s only account memberships.