[#132794] Get rid of unnecessary error and bang method (#788)
* [#132794] Get rid of unnecessary error and bang method

The signature of the `AASM::InvalidTransition` changed in the new
version, but we weren’t hitting this code (because it should be truly
exceptional) so we didn’t know. Then we tried to update a quantity to
zero, and this blew up.

The save! happens shortly after this, so here we only need to do one
save (hence getting rid of the bang).