[#133125] Prevent creation of usernames with extra whitespace
When creating a new external user, any surrounding whitespace should be
ignored when looking up the user as well as when it creates the user.

This had caused a problem for DC where one of their users was unable to
log in, and also because email and username did not match, were unable
to use the forgot password feature.