[#133153] Project not persisting on reservation creation
What was happening is the project_id was being set in the reservation
controller, but then when it redirects to Order#purchase, the
project_id was being reset to NULL.

I discovered this because the newer `.show-for-print` was erring out on
this view.

This also adds feature specs for ordering with a project.