[#133148] Accessories not getting pricing and fulfilled_at set
The symptom as described by NU was “there are 5 orders that are marked
as completed but have not moved to Send Notifications. The color of
their pricing is still the orange yellow and not green”.

Here is a way to reproduce:

* Complete a reservation on an instrument that has accessories
* Find the order on "All Transactions"
* Click into the order (not order detail)
* Click the + to "Add Accessories"
* Add an accessory
* The new order detail does not have a fulfilled_at or get its pricing
assigned. It also is in state "new", but it is in the correct
"Complete" order status.