[#130583] Creates separate print view (#809)
* Creates separate print view

This creates a separate print view for the details normally displ-
ayed in the form for printing. This also removes the now unneeded
styles from the print file

* Organize edit view for print and edit

This organizes the edit view a little more so that there are
partials for print, regular form, and shared details section. This
also adds a view hook for the projects section

* Updates view hook and updates conditionals

This adds some more conditionals where needed and fixes the view
hook that shows the project select on the form

* Hide actual times fields from print view if no actual times