[#133667] Replace TimeCop with ActiveSupport's TimeHelpers (#841)
* The :time_travel rspec metadata tag replaces :timecop_freeze
It works as before, where if let(:now) is defined, it will travel to that time,
run the specs, and then return.

* The clock does not move forward unless moved with a travel() or travel_to()
When changing times with Timecop, it was possible to freeze time in place, or
to have the clock continue to tick forward. Time helpers will only freeze time,
and the entire test suite is run under a time travel, so tests cannot rely on
time moving forward whether they are explicitly using time traveling or not.

* travel()/travel_to() has a block form, but it resets to real time
Time helpers do not support nesting. This change adds two helpers,
travel_and_return and travel_to_and_return, that emulate Timecop's nestable
behavior.