Display percent discount on schedule rule calendar instead of reserve interval
The reserve interval doesn't really make sense anymore now that the setting has
moved from the reserve interval has moved from the schedule rule to the instrument
(way back when in #80). Better to show the percent discount.

This will also make more sense when we genericize schedule rules to apply to
other product types like secure rooms which do not have a reserve interval.

This also adds the scheduling groups to the ScheduleRule#index view. I've intentionally
not done I18n as I plan to do that as part of genericizing this view for secure rooms.

In addition, it does some renaming to make it a presenter, which is closer to what it
actually is.

Add schedule groups to ScheduleRule#index