[#135513] Reservations on split accounts were being deleted
Between Rails 4.1 & 4.2, the behavior of has_one changed. Now, if the model (i.e.) the
order_detail is persisted, calling `order_detail.reservation =` will trigger
a `DELETE FROM reservations where order_detail_id = ?`.

Our `dup`ed order detail was being considered persisted because we copy over
the ID (which we need for reporting).

By moving the ID copying until last, Rails will assume that the order detail
is not persisted when we assign the reservation, so it won't trigger the DELETE.