Fix ability matcher (#948)
Introduced in #931, this is intended to test on a block if it's given, otherwise
use the object. We don't actually use the block version of the positive assertion,
but we do use the negative. This was meant to be parallel to the negative version, but looks like I messed
it up.