[#134079] Revert to using jquery-ui's Smoothness theme (#960)
* [#134079] Revert to using query-ui's Smoothness them

Smoothness was the default pre-packaged theme in jquery-ui 1.11. In
1.12, there’s a new “Base” theme. This was upgraded when we upgraded
jquery-ui-rails to 6.0 in PR #910.

In order to get the old theme, I had to download the theme from
https://jqueryui.com/themeroller and put the CSS and images into the
vendor/assets folder. I also needed to update the relative urls for
some image paths (was `url(“images/xxx”)`, needed to be
`url(“jquery-ui-smoothness/xxx”)`.

There were also some images from an even older theme called “Eggplant”,
which I believe is a previous default theme, so I removed those images.

This should revert the date picker widgets as well as the tabs in
reporting to their old styling.

* Exclude vendor/engines from csslint