Make jQuery-ui theme an ERB so it uses rails `image_tag` helper
We had problems on deployment because the image assets weren't in the hard-coded
paths we expected (including digest) when we updated to revert to the Smoothness
theme of previous jquery-ui versions.

Inspired by:
https://github.com/jquery-ui-rails/jquery-ui-rails/blob/master/app/assets/stylesheets/jquery-ui/theme.css.erb

The jquery-ui-rails repo has a complicated Rakefile that does similar building
from the "base" theme, but simpler instructions have been provided in
`vendor/assets/stylesheets/jquery-ui.theme.md`.