[#134318] Fix Project chosen width on Order Detail popup
Sometimes, when the order detail modal pops up, chosen would not get the width
of the select box.

While a little kludgy, this adds a short timeout before loading
chosen to ensure that the browser has had time to set the box's width before
Chosen looks at it.

This also fixes a minor height issue on the search box where the box was too short.