[#135584] Prevent hard error when trying to purchase an instrument with no reservation
`orders#update_or_purchase (NoMethodError) "undefined method `destroy' for nil:NilClass"`

I'm not 100% how to replicate the error, but it appears that through some multiple
tab situation, there was a "Purchase" button available on the cart, even though
the order did not have a reservation on it. After failing and raising the error,
there was no reservation to destroy.