[#134547] Set fulfilled_at when completing order from modal (#980)
* Show fulfilled at field on popup when marking complete

Refactored the fulfilled at field on the add to order to these are more
similar (using data attributes to show the relationship).

* Set fulfilled_at on backend when submitting from modal

Also, refactor how we do fulfilled_at when adding to an existing order

It sets the fulfilled_at to noon of the day you choose. This matches the behavior of the existing feature.