[#135958] Fix error on item and service access list index (#992)
Prior to #937, we were always setting the route for the `form_for` to
`facility_instrument_update_restrictions_path`, which would route, but
would be invalid. But services and items would never have schedule
groups so the submit would never appear, so we never had any problems.

In #937, we made it generic to support secure rooms, which also support
schedule groups.