[#144059457] Display occupancy times on billing screens (#993)
* Refactor out a TimeRange object

In order to consolidate range rendering and minute calculation that
will also be used by SecureRooms::Occupancy, refactor out a new
TimeRange object that handles these functions.

---

I liked dan's comment about possibly thinking of some of this as presentation logic, but there is such a large reliance right now on Reservation#to_s as well as order_detail.reservation, that I chose to leave it there for now (although I did end up delegating it to a new presenter object), which also includes a new TimeRange object. That new class does mix some presentation (its to_s method) with business logic (duration_mins), and doesn't seem so complicated yet as to warrant splitting, but I could see it wanting to be split in the future.

There are places in business logic (see OrderDetail#set_problem_order) where we will need to switch to using something like order_detail.time_data.

Now that Reservations::Rendering is a more reasonable size, I'm thinking of folding it back into the main Reservation class.

One path I went down was using the TimeData.for mode of attack and have it return a SimpleDelegator around either reservation or occupancy, but try has problems with delegators, and there seem to be enough trys around that it might not be worth it. But something that may be worth revisiting.